# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PORTAL API Documentation",
    "version": "1.0.0",
    "description": "Documentación técnica de la API de PORTAL - Sistema de Gestión de Personal y Horarios."
  },
  "servers": [{ "url": "http://localhost:4000", "description": "Servidor de Desarrollo" }],
  "components": {
    "schemas": {
...
      "UpdateUser": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "minLength": 3 },
          "password": { "type": "string", "minLength": 6 },
          "role": {
            "type": "string",
            "enum": [
              "Usuario",
              "Reloj_Control",
              "Supervisor",
              "Administrador",
              "Supervisor_Elevado",
              "Fiscalizador"
            ]
          },
          "employeeId": { "type": "string" }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "username": { "type": "string" },
          "role": {
            "type": "string",
            "enum": [
              "Usuario",
              "Reloj_Control",
              "Supervisor",
              "Administrador",
              "Supervisor_Elevado",
              "Fiscalizador",
              "Archivado",
              "Reloj Control",
              "Supervisor Elevado"
            ]
          },
          "employeeId": { "type": "string", "nullable": true },
          "mustChangePassword": { "type": "boolean" },
          "mfaEnabled": { "type": "boolean" },
          "lastLogin": { "type": "string", "nullable": true },
          "createdAt": { "type": "string" },
          "updatedAt": { "type": "string" },
          "isDeleted": { "type": "boolean" },
          "syncStatus": { "type": "string" },
          "lastModified": { "type": "number" }
        },
        "required": ["id", "username", "role"],
        "additionalProperties": false
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/api/users": {
      "get": {
        "summary": "Lista todos los usuarios (Admin only)",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "since",
            "schema": { "type": "string" },
            "description": "Timestamp para consulta delta"
          },
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer" },
            "description": "Número de página"
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": { "type": "integer" },
            "description": "Cantidad de items por página"
          },
          {
            "in": "query",
            "name": "search",
            "schema": { "type": "string" },
            "description": "Término de búsqueda (nombre o username)"
          },
          {
            "in": "query",
            "name": "role",
            "schema": { "type": "string" },
            "description": "Filtro por rol"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios o respuesta paginada",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/User" }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "total": { "type": "integer" },
                            "page": { "type": "integer" },
                            "totalPages": { "type": "integer" }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crea un nuevo usuario",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "role"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "role": { "type": "string" },
                  "employeeId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          }
        }
      }
    },
    "/api/users/{id}": {
      "put": {
        "summary": "Actualiza un usuario existente",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateUser" } }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          }
        }
      },
      "delete": {
        "summary": "Elimina un usuario",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "204": { "description": "Usuario eliminado" } }
      }
    },
...
  },
...
}